requirements.txt
streamlit==1.36.0
langchain==0.2.1
langchain-google-genai==0.1.3
pypdf==4.2.0
python-dotenv==1.0.1
faiss-cpu==1.8.0